# server/src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "World of Stories API",
    "version": "1.0.0",
    "description": "Documentación limpia y modularizada"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
  "paths": {
...
    "/users/{id}": {
      "get": {
        "operationId": "getUser",
        "tags": [
          "Users"
        ],
        "summary": "Obtener usuario por ID (getUserById)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID único de MongoDB del usuario",
            "schema": {
              "type": "string",
              "example": "665ed8d5c3f4e567890efabc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado."
          }
        }
      },
      "put": {
        "operationId": "updateUser",
        "tags": [
          "Users"
        ],
        "summary": "Actualizar un usuario completo (updateUser)",
        "description": "Reemplaza o actualiza las propiedades modificables del perfil de usuario mediante ID por parámetro.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userName": {
                    "type": "string",
                    "example": "Kaladin_BridgeFour"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "kaladin@bridgefour.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "kaladin123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontró el usuario para actualizar."
          }
        }
      },
      "delete": {
        "operationId": "deleteUser",
        "tags": [
          "Users"
        ],
        "summary": "Eliminar un usuario (deleteUser)",
        "description": "Remueve físicamente el registro del usuario del sistema utilizando el ID por parámetro.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario eliminado del sistema con éxito.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Usuario eliminado correctamente."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Location": {
        "type": "object",
        "description": "Representa la ubicación exacta del personaje dentro del mapa interactivo",
        "properties": {
          "cityId": {
            "type": "string",
            "nullable": true,
            "example": "665ed7c4c3f4e567890defab"
          },
          "placeId": {
            "type": "string",
            "nullable": true,
            "example": "665ed5a1c3f4e567890bcdef"
          },
          "npcId": {
            "type": "string",
            "nullable": true,
            "example": "665edaa3c3f4e567890f1234"
          },
          "chapterId": {
            "type": "string",
            "nullable": true,
            "example": "665edcc5c3f4e567890f5678"
          }
        }
      },
      "Character": {
        "type": "object",
        "description": "Entidad de personaje que almacena el estado de la partida y el progreso individual",
        "properties": {
          "name": {
            "type": "string",
            "example": "Kaladin"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "money": {
            "type": "number",
            "example": 150
          },
          "items": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "Lanza de entrenamiento",
              "Frasco de esferas"
            ]
          },
          "fragmentsRead": {
            "type": "array",
            "description": "IDs de relatos o fragmentos del Cosmere que este personaje específico ya leyó",
            "items": {
              "type": "string"
            },
            "example": [
              "665ed1b2c3f4e567890abcde"
            ]
          },
          "animations": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "User": {
        "type": "object",
        "required": [
          "email",
          "username",
          "password"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID nativo de MongoDB",
            "example": "665ed8d5c3f4e567890efabc"
          },
          "id": {
            "type": "string",
            "description": "ID casteado para compatibilidad en Frontend",
            "example": "665ed8d5c3f4e567890efabc"
          },
          "username": {
            "type": "string",
            "example": "kaladin31"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "kaladin@bridgefour.com"
          },
          "password": {
            "type": "string",
            "description": "Hash encriptado",
            "example": "$2b$10$X7rE8Y3M..."
          },
          "rol": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "USER_ROLE"
            ]
          },
          "characters": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Character"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-08T14:22:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-08T15:30:00.000Z"
          }
        }
      },
...
    }
  }
}